pyproject.toml
[project]
name = "sermon2jekyllmd"
version = "0.1.0"
description = ""
authors = [
    {name = "Ken",email = "tk.fang@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-docx (>=1.2.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "sermon2jekyllmd", from = "src"}]

[tool.poetry.scripts]
sermon2jekyllmd = "sermon2jekyllmd.cli:app"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
